refactor: handleFilterChange 이벤트 핸들러를 Filters 컴포넌트로 이동
