+++ src/components/Filters.js
@@ -1,8 +1,8 @@
-import { CATEGORY, SORT } from '../constants';
+import { CATEGORY_OPTIONS, SORT_OPTIONS } from '../constants';
 import selectTemplate from '../template/selectTemplate';
 import { arrayElementToObject } from '../utils/util';
 
-export default function Filters($root, handleFiltersChange) {
+export default function Filters($root) {
   const $filterSection = document.createElement('section');
   $filterSection.className = 'restaurant-filter-container';
 
@@ -19,21 +19,21 @@ export default function Filters($root, handleFiltersChange) {
 
   this.render = () => {
     $filterSection.innerHTML = `
-    ${selectTemplate({
-      name: 'category',
-      id: 'category-filter',
-      options: arrayElementToObject(['전체', ...CATEGORY]),
-      selected: this.state.category,
-      className: 'restaurant-filter',
-    })}
-    ${selectTemplate({
-      name: 'sorting',
-      id: 'sorting-filter',
-      options: arrayElementToObject(SORT),
-      selected: this.state.filter,
-      className: 'restaurant-filter',
-    })}
-  `;
+      ${selectTemplate({
+        name: 'category',
+        id: 'category-filter',
+        options: arrayElementToObject(['전체', ...CATEGORY_OPTIONS]),
+        selected: this.state.category,
+        className: 'restaurant-filter',
+      })}
+      ${selectTemplate({
+        name: 'sorting',
+        id: 'sorting-filter',
+        options: arrayElementToObject(SORT_OPTIONS),
+        selected: this.state.filter,
+        className: 'restaurant-filter',
+      })}
+    `;
   };
 
   this.setState = (state) => {
@@ -43,3 +43,19 @@ export default function Filters($root, handleFiltersChange) {
 
   this.init();
 }
+
+const handleFiltersChange = (event) => {
+  const { filters } = this.state;
+  const { id, value } = event.target;
+
+  switch (id) {
+    case 'category-filter':
+      filterRestaurantList(value, filters.state.filter);
+      break;
+    case 'sorting-filter':
+      filterRestaurantList(filters.state.category, value);
+      break;
+    default:
+      return;
+  }
+};
